rdb-gatewayの不要なmain.rs削除とrepositories/user.rsへの分割
